Extensions: detect remote repositories when dropping a file
When dropping an extension into Blender, check if the extensions is
part of a remote repository which is then used to set the default
repository.

This makes it more convenient to download larger extensions from
remote repositories using a web-browser & drop them into Blender after,
rather than dropping the URL directly into Blender.