VSE: Fix linked time option for sequencer select
When 7afcfe1 removed the use of `startdisp` and `enddisp` for everything
but effect strips, not all of these variables were replaced in
`select_linked_time`, breaking the option. This option is used for the
select operator with the ctrl modifier in both LCS and RCS default
keymaps.

This patch fixes the bug, and also cleans up the old C code in the
function, replacing it with more robust logic.
- The new logic allows for the "linked time" option to be combined with
  the "toggle" option -- old logic only propagated deselects if both
  left and right handles were aligned.
- New logic makes sure that a selection is only propagated along the
  side that you click by adding the selection handle as an argument.
  (Before, you could align two strips on the left side only, "regular" select the
  left handle of the top one, then "linked time" select the right
  handle, and it would erroneously propagate the left handle selection
  too).

This patch also fixes a bug where "both handle" selection would not work
if the linked time option was set, by making sure that if `seq2` is set
in `StripSelection`, then `select_linked_time` is run once more.

Pull Request: https://projects.blender.org/blender/blender/pulls/125039